requirements.txt
scrapegraphai
playwright
ddgs
requests
python-dotenv
